Updates to controllers, app.js and routes. Addition of js pages

+++ server/controllers/conversationsController.js
@@ -2,19 +2,13 @@
 const Conversation = require('../models/conversation');
 const User = require('../models/user');
 
-exports.new = (req, res) => {
-    req.isAuthenticated();
-    res.render('conversations/new', {
-      title: 'New Conversation'
-    });
-  };
-
 exports.create = (req, res) => {
-    req.isAuthenticated();
+    if (!req.isAuthenticated()) 
+      return res.status(401).send({error: "Not Authenticated"});
     //Find the recipient in the database
     User.findOne({email: req.body.email})
         .then(recipient => {
-            return recipient._id;
+            return recipient._id; 
         })
         .then((recipientId => {
             //Create a new conversation whose owners are the current session user and recipient
@@ -22,73 +16,53 @@ exports.create = (req, res) => {
                 users: [req.session.userId, recipientId],
                 message: [] //empty conversation
             }))
-                .then(()=> {
-                    req.flash('success', "Conversation created successfully");
-                    res.redirect('/conversations'); //should probably open the new conversation instead
-                })
-                .catch(err => {
-                    req.flash('err', `${err}`);
-                    res.redirect('/conversations/new');
-                });
+                .then(()=> res.status(200).send({ success: "Conversation created successfully" }))
+                .catch(err => res.status(404).send(err));
     }))
-        .catch(err => {
-            req.flash('error', `Could not find the recipient: ${err}`);
-            res.redirect('/conversations/new');
-    });
+        .catch(err => res.status(404).send({error: "Could not find the recipient" }));
 }
 
 exports.index = (req, res) => {
-    req.isAuthenticated();
+    if (!req.isAuthenticated()) 
+    	return res.status(401).send({error: "Not Authenticated"});
+
     Conversation.find({
         //conversations in which the current user is a participant
         users: {$elemMatch: {$in: [req.session.userId]}}
     })
         .populate('users')
-        .then((conversations) => {
-            res.render('conversations/index', {
-                conversations: conversations,
-                title: "Conversations"
-            });
-        })
-        .catch(err => {
-            req.flash('error', `Error finding conversations: ${err}`);
-            res.redirect('/');
-        });
+		.then((conversations) => res.json(conversations))
+        .catch(err => res.status(404).json(err));
 }
 
 //shows the contents of the conversation - all the messages. Actually uses the messages/index view
 exports.show = (req, res) => {
-    req.isAuthenticated();
+    if (!req.isAuthenticated()) 
+    	return res.status(401).send({error: "Not Authenticated"});
     //receives a conversation id. We need to display the correct message
     Conversation.findById(req.params.id)
         .populate('messages.user')
         .populate('users')
-        .then(conversation => {
-            req.flash('success', "Found the conversation");
-            res.render('messages/index', {
-                title: "Talk about it",
-                conversation: conversation,
-                sessionUserId: req.session.userId
-            });
-        })
-        .catch(err => {
-            req.flash('error', `${err}`);
-            res.redirect('/conversations');
-        })
+		.then(conversation => res.json(conversation))
+					// {
+          //   req.flash('success', "Found the conversation");
+          //   res.render('messages/index', {
+          //       title: "Talk about it",
+          //       conversation: conversation,
+          //       sessionUserId: req.session.userId
+					// 	}
+						// );
+        // })
+        .catch(err => res.status(404).json(err));
 }
 
 exports.destroy = (req, res) => {
-    req.isAuthenticated();
+    if (!req.isAuthenticated()) 
+    	return res.status(401).send({error: "Not Authenticated"});
     Conversation.deleteOne({
         _id: req.body.id
     })
-        .then(() => {
-            req.flash('success', "Conversation deleted");
-            res.redirect('/conversations');
-        })
-        .catch(err => {
-            req.flash('error', `Error: ${err}`);
-            res.redirect('/conversations');
-        });
+        .then(() => res.status(200).send({ success: "Conversation deleted" }))
+        .catch(err => res.status(404).send(err));
 }
